% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_inference.R
\name{get_asymp_var_rho}
\alias{get_asymp_var_rho}
\title{Get Asymptotic Variance for Rho}
\usage{
get_asymp_var_rho(theta, region1_info, region2_info)
}
\arguments{
\item{theta}{Numeric. The estimated stage 2 parameter vector}

\item{region1_info}{Stage 1 info for region 1.}

\item{region2_info}{Stage 1 info for region 2.}
}
\value{
A scalar value representing the asymptotic variance.
}
\description{
Computes the asymptotic variance for rho based on the Fisher information matrix.
}
