% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_cor_mat}
\alias{get_cor_mat}
\title{Compute correlation matrix}
\usage{
get_cor_mat(corfun_name_string, xSqrd_mat, tau)
}
\arguments{
\item{corfun_name_string}{Choice of kernel function}

\item{xSqrd_mat}{Matrix of squared difference between inputs x's}

\item{tau}{Inverse lengthscale parameter}
}
\value{
cormat Correlation matrix from vector of inputs x's
}
\description{
Compute correlation matrix
}
\details{
The correlation matrix with corresponding choice of kernel: For RBF kernel \deqn{C(x_i,x_j) = \exp(-\tau^2/2 \times d^2),}
for Matern 1/2 kernel \deqn{C(x_i,x_j) = \exp(-\tau \times d),}
for Matern 3/2 kernel \deqn{C(x_i,x_j) = (1 + \sqrt{3} \times\tau \times d)\exp(-\sqrt{3}\times\tau \times d),}
and for Matern 5/2 kernel \deqn{C(x_i,x_j) = (1 + \sqrt{3} \times\tau \times d + \sqrt{5}/\sqrt{3} \times\tau^2 \times d^2)\exp(-\sqrt{5}\times\tau \times d),}
with \eqn{d = || x_i - x_j ||_2^2} is the squared distance between locations i and j.
}
