% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_dist_sqrd_mat}
\alias{get_dist_sqrd_mat}
\title{Get distance squared matrix for simulated data}
\usage{
get_dist_sqrd_mat(L, sideLength, voxelID)
}
\arguments{
\item{L}{Number of voxels}

\item{sideLength}{side length of that region}

\item{voxelID}{ID of sampled voxels}
}
\value{
Matrix of squared distance between voxels for that region
}
\description{
Get distance squared matrix for simulated data
}
