% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_dist_sqrd_mat_from_coord}
\alias{get_dist_sqrd_mat_from_coord}
\title{Get distance squared matrix for real data}
\usage{
get_dist_sqrd_mat_from_coord(coord_mat)
}
\arguments{
\item{coord_mat}{Matrix with each column is coordinate of each voxel}
}
\value{
Matrix of squared distance between voxels for that region
}
\description{
Get distance squared matrix for real data
}
