% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kronecker_mvm}
\alias{kronecker_mvm}
\title{Kronecker matrix-vector product (Steeb & Hardy, 2011)}
\usage{
kronecker_mvm(A, B, v)
}
\arguments{
\item{A}{Left matrix}

\item{B}{Right matrix}

\item{v}{vector}
}
\value{
Kronecker matrix-vector product
}
\description{
\deqn{(A \otimes B)v = \text{vectorise}(BVA^\top)}
}
\references{
Willi-Hans Steeb and Yorick Hardy. Matrix calculus and Kronecker product: a practical approach to linear and multilinear algebra. World Scientific Publishing Company, 2011.
}
