% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{opt_inter}
\alias{opt_inter}
\title{Fit inter-regional model using L-BFGS}
\usage{
opt_inter(
  theta_init,
  X,
  Z,
  dist_sqrdMat_1,
  dist_sqrdMat_2,
  time_sqrd_mat,
  gamma_vec,
  kernel_type
)
}
\arguments{
\item{theta_init}{unrestricted initialization of parameters  for inter-regional model}

\item{X}{Data matrix of signals of 2 regions}

\item{Z}{fixed-effects design matrix of 2 regions}

\item{dist_sqrdMat_1}{Block component for that region 1}

\item{dist_sqrdMat_2}{Block component for that region 2}

\item{kernel_type}{Choice of spatial kernel}
}
\value{
List of 3 components:
\item{theta}{estimated inter-regional parameters}
\item{asymptotic_var}{asymptotic variance of transformed correlation coefficient}
\item{rho_transformed}{Fisher transformation of correlation coefficient}
}
\description{
Fit inter-regional model using L-BFGS
}
