% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{opt_intra}
\alias{opt_intra}
\title{Fit intra-regional model using L-BFGS}
\usage{
opt_intra(
  theta_init,
  X_region,
  Z_region,
  dist_sqrd_mat,
  time_sqrd_mat,
  L,
  M,
  kernel_type
)
}
\arguments{
\item{theta_init}{unrestricted initialization of parameters for 1 region}

\item{X_region}{Data matrix of signals of 1 region}

\item{Z_region}{fixed-effects design matrix of 1 region}

\item{dist_sqrd_mat}{Spatial squared distance matrix}

\item{time_sqrd_mat}{Temporal squared distance matrix}

\item{L}{Number of voxels}

\item{M}{Number of time points}

\item{kernel_type}{Choice of spatial kernel}
}
\value{
List of 2 components:
\item{theta}{estimated intra-regional parameters}
\item{nu}{fixed-effect estimate}
}
\description{
Fit intra-regional model using L-BFGS
}
