% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncMM_fullrun_small_regions.R
\name{qfuncMM_fullrun_small_regions}
\alias{qfuncMM_fullrun_small_regions}
\title{Run both stage 1 and stage 2 for QFuncMM. This is a convenience function for small brain regions and testing.
Larger applications should use `qfuncMM_stage1_intra` and `qfuncMM_stage2_inter` separately.}
\usage{
qfuncMM_fullrun_small_regions(
  region_list,
  voxel_coords,
  kernel_type = "matern_5_2",
  stage1_cov_setting = c("noisy", "noiseless"),
  num_init = 10L,
  verbose = FALSE
)
}
\arguments{
\item{region_list}{List of \eqn{M\times L_j} region matrices.}

\item{voxel_coords}{List specifying voxels for each region. Each item
in the list is a \eqn{L_j \times 3} matrix of spatial coordinates.}

\item{kernel_type}{Choice of spatial kernel.}

\item{stage1_cov_setting}{Choice of noisy or noiseless stage 1 model.}

\item{num_init}{Number of stage 1 initializations to use.}

\item{verbose}{Print progress messages.}
}
\description{
Run both stage 1 and stage 2 for QFuncMM. This is a convenience function for small brain regions and testing.
Larger applications should use `qfuncMM_stage1_intra` and `qfuncMM_stage2_inter` separately.
}
