% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncMM_stage1_intra.R
\name{qfuncMM_stage1_intra}
\alias{qfuncMM_stage1_intra}
\title{Estimate functional connectivity from voxel-level BOLD signals. Run stage 1 intra-regional analysis.
  Results are saved to a JSON file.}
\usage{
qfuncMM_stage1_intra(
  region_list,
  voxel_coords,
  kernel_type = "matern_5_2",
  cov_setting = c("standard", "diag_time", "noiseless", "noiseless_profiled"),
  out_file = ".",
  overwrite = FALSE,
  num_init = 10L,
  verbose = TRUE
)
}
\arguments{
\item{region_list}{List of \eqn{M\times L_j} region matrices.}

\item{voxel_coords}{List specifying voxels for each region. Each item
in the list is a \eqn{L_j \times 3} matrix of spatial coordinates.}

\item{kernel_type}{Choice of spatial kernel.}

\item{cov_setting}{Choice of covariance structure.}

\item{out_file}{Output file path.}

\item{overwrite}{Overwrite output file if it exists.}

\item{num_init}{Number of initializations for multi-start optimiization.}

\item{verbose}{Print progress messages.}
}
\description{
Estimate functional connectivity from voxel-level BOLD signals. Run stage 1 intra-regional analysis.
  Results are saved to a JSON file.
}
