% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncMM_stage2_vecchia.R
\name{qfuncMM_stage2_vecchia}
\alias{qfuncMM_stage2_vecchia}
\title{Estimate functional connectivity from voxel-level BOLD signals.
  Run stage 2 inter-regional analysis for a pair of regions using Vecchia's approximation.
  Results are saved to a JSON file.}
\usage{
qfuncMM_stage2_vecchia(
  stage1_region1_outfile,
  stage1_region2_outfile,
  out_dir,
  m_seq = 100,
  st_scale = c(10, 1),
  data_and_coords = NULL,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{stage1_region1_outfile}{JSON file from stage 1 for region 1.}

\item{stage1_region2_outfile}{JSON file from stage 1 for region 2.}

\item{out_dir}{Output directory. If NULL, no JSON file is written.}

\item{m_seq}{Number of neighbors to use in Vecchia approximation.}

\item{st_scale}{Scaling factor for spatial and temporal dimensions.}

\item{data_and_coords}{Optional list containing standardized data and coordinates for both regions,
if not provided in stage 1 files.}

\item{overwrite}{Overwrite existing output file.}

\item{verbose}{Print progress messages.}
}
\value{
A list containing the results of the inter-regional analysis.
}
\description{
Estimate functional connectivity from voxel-level BOLD signals.
  Run stage 2 inter-regional analysis for a pair of regions using Vecchia's approximation.
  Results are saved to a JSON file.
}
