% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_3_region.R
\name{simulate_3_region}
\alias{simulate_3_region}
\title{Simulate signals from 3 regions}
\usage{
simulate_3_region(
  L1,
  L2,
  L3,
  side_length,
  M,
  theta,
  rho_vec,
  sigma_sqrd = 1,
  mu_1,
  mu_2,
  mu_3,
  vxlID_1,
  vxlID_2,
  vxlID_3,
  random_seed,
  num_sim,
  out_folder_name,
  plot_it = T,
  C_kernel_type = "matern_5_2"
)
}
\arguments{
\item{L1}{number of voxels in region 1}

\item{L2}{number of voxels in region 2}

\item{L3}{number of voxels in region 3}

\item{side_length}{side length of 3D lattice where voxels are sampled}

\item{M}{number of timepoints}

\item{theta}{true covariance matrices parameters}

\item{rho_vec}{vector of inter-regional correlations \eqn{(\rho_{12}, \rho_{13}, \rho_{23})}}

\item{sigma_sqrd}{noise variance. Default 1.}

\item{mu_1}{\code{M}-by-\code{1} vector of mean of signal from region 1}

\item{mu_2}{\code{M}-by-\code{1} vector of mean of signal from region 2}

\item{mu_3}{\code{M}-by-\code{1} vector of mean of signal from region 3}

\item{vxlID_1}{\code{L_1}-by-\code{1} vector of voxel ID of region 1}

\item{vxlID_2}{\code{L_2}-by-\code{1} vector of voxel ID of region 2}

\item{vxlID_3}{\code{L_3}-by-\code{1} vector of voxel ID of region 3}

\item{random_seed}{random seed}

\item{num_sim}{number of simulation}

\item{out_folder_name}{folder to save simulated data}

\item{plot_it}{If True, plot of simulated data is saved in \code{out_folder_name}. Default True.}

\item{C_kernel_type}{Choice of spatial kernel. Defaul "matern_5_2".}
}
\value{
obs_signal Simulated signal
}
\description{
Simulate signals from 3 regions
}
\examples{
L <- 20 # Numbers of voxels
side_length <- 7 # Side length of each region
M <- 30 # Numbers of timepoints
num_sim <- 1 # Numbers of simulation
# Generate voxels for each region
set.seed(1)
vxlID_1 <- sample(1:(side_length^3), L, replace = FALSE)
set.seed(2)
vxlID_2 <- sample(1:(side_length^3), L, replace = FALSE)
set.seed(4)
vxlID_3 <- sample(1:(side_length^3), L, replace = FALSE)

# rho
rho_vec <- data.frame(rho12=0.1, rho13=0.35, rho23=0.6)
# Parameters
parameters_true <- data.frame(tau_eta = 1/4, k_eta = 0.5,
phi_gamma_1=1, tau_gamma_1=1/2, k_gamma_1=2,
phi_gamma_2=1, tau_gamma_2=1/2, k_gamma_2=2,
phi_gamma_3=1, tau_gamma_3=1/2, k_gamma_3=2,
nugget=0.1, mu_1 = 1, mu_2 = 10, mu_3 = 20)

simulated_data = simulate_3_region(L1=L, L2=L, L3=L,
                                   side_length=side_length, 
                                   M=M, 
                                   theta=parameters_true, 
                                   rho_vec=rho_vec, 
                                   sigma_sqrd=1,
                                   mu_1=rep(parameters_true$mu_1, M),
                                   mu_2=rep(parameters_true$mu_2, M),
                                   mu_3=rep(parameters_true$mu_3, M),
                                   vxlID_1=vxlID_1, 
                                   vxlID_2=vxlID_2, 
                                   vxlID_3=vxlID_3,
                                   random_seed=1, num_sim=1,
                                   C_kernel_type="matern_5_2")
plot_mat <- matrix(c(simulated_data[[1]]$X_Region1, simulated_data[[1]]$X_Region2, simulated_data[[1]]$X_Region3), 3*L, M, byrow=TRUE)
matplot(t(plot_mat), type="l", xlab="Time points", ylab="Signal")
}
