% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_inference.R
\name{get_asymp_var_rho}
\alias{get_asymp_var_rho}
\title{Get Asymptotic Variance for Rho}
\usage{
get_asymp_var_rho(
  theta,
  region1_info,
  region2_info,
  method = c("reml", "vecchia", "vecchia_diag_time")
)
}
\arguments{
\item{theta}{Numeric. The estimated stage 2 parameter vector}

\item{region1_info}{Stage 1 info for region 1.}

\item{region2_info}{Stage 1 info for region 2.}

\item{method}{Which method was used in stage 2? Either `reml`, `vecchia`, or `vecchia_diag_time`.
Option `vecchia_diag_time` is a faster approximation of the Fisher information matrix by using the diagonal of the temporal covariance matrix.}
}
\value{
A scalar value representing the asymptotic variance.
}
\description{
Computes the asymptotic variance for rho based on the Fisher information matrix.
}
