% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncMM.R
\name{qfuncMM}
\alias{qfuncMM}
\title{Estimate functional connectivity from voxel-level BOLD signals.}
\usage{
qfuncMM(
  region_list,
  voxel_coords,
  kernel_type = "matern_5_2",
  cov_setting = c("standard", "diag_time", "noiseless", "noiseless_profiled"),
  verbose = TRUE
)
}
\arguments{
\item{region_list}{List of \eqn{M\times L_j} region matrices.}

\item{voxel_coords}{List specifying voxels for each region. Each item
in the list is a \eqn{L_j \times 3} matrix of spatial coordinates.}

\item{kernel_type}{Choice of spatial kernel. Default "matern_5_2".}

\item{cov_setting}{Choice of covariance setting.}

\item{verbose}{Print progress messages.}
}
\description{
Estimate functional connectivity from voxel-level BOLD signals.
}
