% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncMM_stage1_intra.R
\name{qfuncMM_stage1_intra}
\alias{qfuncMM_stage1_intra}
\title{Estimate functional connectivity from voxel-level BOLD signals. Run stage 1 intra-regional analysis for a single region.
  Results are saved to a JSON file.}
\usage{
qfuncMM_stage1_intra(
  subject_id,
  region_uniqid,
  region_name,
  region_data,
  region_coords,
  out_dir,
  kernel_type = "matern_5_2",
  cov_setting = c("auto", "noisy", "noiseless"),
  num_init = 10L,
  verbose = FALSE
)
}
\arguments{
\item{subject_id}{Identifies the subject/exam.}

\item{region_name}{Name of the region.}

\item{region_data}{\eqn{M\times L_j} matrix}

\item{region_coords}{\eqn{L_j \times 3} matrix of spatial coordinates.}

\item{out_dir}{Output directory.}

\item{kernel_type}{Choice of spatial kernel.}

\item{cov_setting}{Choice of covariance structure. Default option 'auto' chooses between 'noisy' and 'noiseless' based on model fit.}

\item{num_init}{Number of initializations for multi-start optimiization.}

\item{verbose}{Print progress messages.}

\item{region_uniqud}{Uniquely identifies the region for a given subject.}
}
\description{
Estimate functional connectivity from voxel-level BOLD signals. Run stage 1 intra-regional analysis for a single region.
  Results are saved to a JSON file.
}
