% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncMM_stage2_inter.R
\name{qfuncMM_stage2_inter}
\alias{qfuncMM_stage2_inter}
\title{Estimate functional connectivity from voxel-level BOLD signals. Run stage 2 inter-regional analysis for a pair of regions.
  Results are saved to a JSON file.}
\usage{
qfuncMM_stage2_inter(
  stage1_region1_outfile,
  stage1_region2_outfile,
  out_dir,
  kernel_type = "matern_5_2",
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{stage1_region1_outfile}{JSON file from stage 1 for region 1.}

\item{stage1_region2_outfile}{JSON file from stage 1 for region 2.}

\item{out_dir}{Output directory.}

\item{kernel_type}{Choice of spatial kernel.}

\item{overwrite}{Overwrite existing output file.}

\item{verbose}{Print progress messages.}
}
\description{
Estimate functional connectivity from voxel-level BOLD signals. Run stage 2 inter-regional analysis for a pair of regions.
  Results are saved to a JSON file.
}
